Move image manipulation tests to its external package
