+++ go.mod
@@ -2,9 +2,6 @@ module github.com/elazarl/goproxy
 
 go 1.20
 
-require (
-	github.com/elazarl/goproxy/ext v0.0.0-20241217120900-7711dfa3811c
-	golang.org/x/net v0.33.0
-)
+require golang.org/x/net v0.33.0
 
 require golang.org/x/text v0.21.0 // indirect
